feat: extract domain types
