+++ src/hooks/__tests__/useEventDetails.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect, vi, beforeEach } from 'vitest';
+import { describe, it, expect, vi, beforeEach, type Mock } from 'vitest';
 import { renderHook, waitFor } from '@testing-library/react';
 import { useEventDetails } from '../useEventDetails';
 import {
@@ -25,16 +25,16 @@ beforeEach(() => {
 
 describe('useEventDetails', () => {
   it('aggregates data from services', async () => {
-    (fetchEvent as any).mockResolvedValue({
+    (fetchEvent as Mock).mockResolvedValue({
       id: '1',
       name: 'Event',
       event_date: '2024-01-01',
       key_info_content: 'info',
     });
-    (fetchPasses as any).mockResolvedValue([
+    (fetchPasses as Mock).mockResolvedValue([
       { id: 'p1', name: 'Pass 1', price: 10, description: 'desc', initial_stock: 10, remaining_stock: 5 },
     ]);
-    (fetchEventActivities as any).mockResolvedValue([]);
+    (fetchEventActivities as Mock).mockResolvedValue([]);
 
     const { result } = renderHook(() => useEventDetails('1'));
 
@@ -45,7 +45,7 @@ describe('useEventDetails', () => {
   });
 
   it('handles errors from services', async () => {
-    (fetchEvent as any).mockRejectedValue(new Error('fail'));
+    (fetchEvent as Mock).mockRejectedValue(new Error('fail'));
     const { result } = renderHook(() => useEventDetails('1'));
     await waitFor(() => expect(result.current.loading).toBe(false));
     expect(result.current.error).toBeTruthy();
